update service test and user

+++ src/modules/user/user.service.ts
@@ -10,10 +10,15 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import * as bcrypt from 'bcrypt';
 import { UpdateStatusDto } from './dto/update-status.dto';
+import { AdminCreateUserDto } from './dto/admin-create-user';
+import { RecaptchaService } from '../auth/recaptcha.service';
 
 @Injectable()
 export class UserService {
-  constructor(@InjectModel(User.name) private userModel: Model<User>) {}
+  constructor(
+    @InjectModel(User.name) private userModel: Model<User>,
+    private readonly recaptchaService: RecaptchaService,
+  ) {}
 
   async findById(id: string): Promise<User> {
     const user = await this.userModel.findById(id);
@@ -28,6 +33,12 @@ export class UserService {
     message: string;
     data: User;
   }> {
+    const isHuman = await this.recaptchaService.verify(
+      createUserDto.recaptchaToken,
+    );
+    if (!isHuman) {
+      throw new BadRequestException('Xác minh không hợp lệ');
+    }
     const { email, password, name } = createUserDto;
     // Check if the email is exist
     const isEmailExist = await this.userModel.exists({ email: email });
@@ -62,26 +73,54 @@ export class UserService {
     return this.userModel.findById(id);
   }
 
-  update(id: string, updateUserDto: UpdateUserDto) {
-    const updateUser = this.userModel.findByIdAndUpdate(
-      id,
-      { ...updateUserDto },
-      { new: true },
-    );
-    return {
-      success: true,
-      message: 'Cập nhật tài khoản thành công',
-      data: updateUser,
-    };
+  async update(id: string, updateUserDto: UpdateUserDto) {
+    try {
+      const user = await this.userModel.findById(id);
+      if (!user) {
+        throw new NotFoundException('Không tìm thấy người dùng');
+      }
+
+      // Nếu có cập nhật mật khẩu
+      if (updateUserDto.password) {
+        const hashedPassword = await bcrypt.hash(updateUserDto.password, 10);
+        updateUserDto.password = hashedPassword;
+      }
+
+      const updateUser = await this.userModel.findByIdAndUpdate(
+        id,
+        { ...updateUserDto },
+        { new: true },
+      );
+
+      return {
+        success: true,
+        message: 'Cập nhật tài khoản thành công',
+        data: updateUser,
+      };
+    } catch (error) {
+      throw new BadRequestException(
+        'Không thể cập nhật thông tin: ' + error.message,
+      );
+    }
   }
 
-  remove(id: string) {
-    const deleteUser = this.userModel.findByIdAndDelete(id);
-    return {
-      success: true,
-      message: 'Xóa tài khoản thành công',
-      data: deleteUser,
-    };
+  async remove(id: string) {
+    try {
+      const user = await this.userModel.findById(id);
+      if (!user) {
+        throw new NotFoundException('Không tìm thấy người dùng');
+      }
+      const deleteUser = await this.userModel.findByIdAndDelete(id);
+      return {
+        success: true,
+        message: 'Xóa tài khoản thành công',
+        data: deleteUser,
+      };
+    } catch (error) {
+      throw new BadRequestException(
+        'Không thể xóa người dùng: ' + error.message,
+      );
+    }
   }
 
   async updateStatus(id: string, updateStatusDto: UpdateStatusDto) {
@@ -98,5 +137,41 @@ export class UserService {
       message: 'Cập nhật trạng thái thành công',
       data: updatedUser,
     };
+  }
+
+  async createByAdmin(adminCreateUserDto: AdminCreateUserDto): Promise<{
+    success: boolean;
+    message: string;
+    data: User;
+  }> {
+    const { email, password, name, role, isApproved, hasAttemptedLogin } =
+      adminCreateUserDto;
+    // Check if the email is exist
+    const isEmailExist = await this.userModel.exists({ email: email });
+    if (isEmailExist) {
+      throw new BadRequestException(
+        'Email đã tồn tại, vui lòng thử lại với email khác',
+      );
+    }
+
+    // Hash password
+    try {
+      const hashedPassword = await bcrypt.hash(password, 10);
+      const user = await this.userModel.create({
+        name,
+        email,
+        password: hashedPassword,
+        role,
+        isApproved,
+        hasAttemptedLogin,
+      });
+      return {
+        success: true,
+        message: 'Tạo tài khoản thành công',
+        data: user,
+      };
+    } catch (error) {
+      throw new BadRequestException(error);
+    }
   }
 }
